Add arch-dev and work VMs
I definitely should have split this into multiple commits.

Added arch-dev - I want to get to the point where I can provision my
arch desktop setup with Ansible, so this is the beginning of that.
Refactored base tasks to be platform-agnostic, and provided a way to run
tasks on different OSes.

Added work_servers for work VM - this is a Debian server with KDE
Plasma, for getting general work-related tasks done. Equipped witha VPN
client for tasks that require it.