Fix git repo cloning
Before, git repos were cloned via HTTP, which wouldn't have worked seamlessly
for private repos. Now, they're cloned via SSH, which requires SSH agent
forwarding to be enabled. However, the socket on the remote system is in a
directory that restricts its access to root users. My solution is to
temporarly change ownership to the user we're becoming, use the socket, and
then change it back immediately after.